kometa/tools/asset-organizer/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kometa-assets-organizer"
description = "Organize and manage Kometa/Plex metadata assets"
dynamic = ["version"]
readme = "README.md"
license = "MIT"
authors = [
    {name = "okdesign21"}
]
keywords = ["kometa", "plex", "assets", "organizer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/okdesign21/ortflix-costume"
Repository = "https://github.com/okdesign21/ortflix-costume"
Documentation = "https://github.com/okdesign21/ortflix-costume/tree/master/kometa/tools/asset-organizer"
Issues = "https://github.com/okdesign21/ortflix-costume/issues"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pyyaml>=6.0",
]

[tool.setuptools]
py-modules = ["organize_assets", "handlers", "poster_handler"]

[tool.setuptools_scm]
# Version derived from git tags (e.g., v0.1.0 -> 0.1.0)
write_to = "_version.py"
local_scheme = "no-local-version"
fallback_version = "0.1.0"
root = "../.."
